refactor(structure): split dedicated spsc core
